+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/panels/PlayButtonController.java
@@ -4,116 +4,128 @@ import de.mephisto.vpin.commons.fx.ConfirmationResult;
 import de.mephisto.vpin.commons.utils.WidgetFactory;
 import de.mephisto.vpin.restclient.PreferenceNames;
 import de.mephisto.vpin.restclient.emulators.GameEmulatorRepresentation;
+import de.mephisto.vpin.restclient.emulators.LaunchConfiguration;
 import de.mephisto.vpin.restclient.frontend.Frontend;
 import de.mephisto.vpin.restclient.frontend.FrontendType;
 import de.mephisto.vpin.restclient.games.GameRepresentation;
 import de.mephisto.vpin.restclient.preferences.UISettings;
 import de.mephisto.vpin.ui.util.FrontendUtil;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
+import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.MenuItem;
-import javafx.scene.control.SeparatorMenuItem;
-import javafx.scene.control.SplitMenuButton;
+import javafx.scene.control.Button;
+import javafx.scene.control.ComboBox;
+import javafx.scene.layout.HBox;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.ResourceBundle;
 
 import static de.mephisto.vpin.ui.Studio.client;
 import static de.mephisto.vpin.ui.Studio.stage;
 
-public class PlayButtonController implements Initializable {
+public class PlayButtonController implements Initializable, ChangeListener<LaunchConfiguration> {
   private final static Logger LOG = LoggerFactory.getLogger(PlayButtonController.class);
 
   @FXML
-  private SplitMenuButton playBtn;
+  private ComboBox<LaunchConfiguration> launchCombo;
+
+  @FXML
+  private Button launchBtn;
+
+  @FXML
+  private HBox root;
 
   private GameRepresentation game;
 
   @Override
   public void initialize(URL location, ResourceBundle resources) {
-    this.playBtn.managedProperty().bindBidirectional(this.playBtn.visibleProperty());
-    this.playBtn.setDisable(true);
+    this.root.managedProperty().bindBidirectional(this.root.visibleProperty());
   }
 
   public void setVisible(boolean b) {
-    this.playBtn.setVisible(b);
+    this.root.setVisible(b);
   }
 
   @FXML
   public void onPlay() {
-    onPlay(null, null);
+    LaunchConfiguration value = launchCombo.getValue();
+    if (value != null) {
+      if (value.isLaunchViaFrontend()) {
+        new Thread(() -> {
+          client.getFrontendService().launchGame(game.getId());
+        }).start();
+      }
+      else {
+        onPlay(game, value.getAltExe(), value.getOption());
+      }
+    }
+
   }
 
   public void setData(GameRepresentation game) {
     this.game = game;
-    playBtn.getItems().clear();
-    playBtn.setDisable(game == null || game.getGameFilePath() == null);
+    boolean disable = game == null || game.getGameFilePath() == null;
+    launchCombo.setDisable(disable);
+    launchBtn.setDisable(disable);
+    this.launchCombo.getItems().clear();
 
-    if (playBtn.isDisabled()) {
+    if (disable) {
       return;
     }
 
+    List<LaunchConfiguration> items = new ArrayList<>();
+    GameEmulatorRepresentation gameEmulator = client.getEmulatorService().getGameEmulator(game.getEmulatorId());
+    String exeName = gameEmulator.getExeName();
+    if (!StringUtils.isEmpty(exeName)) {
+      items.add(new LaunchConfiguration("Emulator Default", false, exeName, null));
+    }
+
     if (client.getFrontendService().getFrontendCached().getFrontendType().equals(FrontendType.Popper)) {
-      MenuItem item = new MenuItem("Launch via PinUP Popper");
-      item.setOnAction(new EventHandler<ActionEvent>() {
-        @Override
-        public void handle(ActionEvent event) {
-          if (game != null) {
-            new Thread(() -> {
-              client.getFrontendService().launchGame(game.getId());
-            }).start();
-          }
-        }
-      });
-      playBtn.getItems().add(item);
-      playBtn.getItems().add(new SeparatorMenuItem());
+      items.add(new LaunchConfiguration("Launch via PinUP Popper", true, null, null));
     }
 
     if (game.isVpxGame()) {
-      GameEmulatorRepresentation gameEmulator = client.getEmulatorService().getGameEmulator(game.getEmulatorId());
+      gameEmulator = client.getEmulatorService().getGameEmulator(game.getEmulatorId());
       List<String> altExeNames = client.getEmulatorService().getAltExeNames(gameEmulator.getId());
       for (String altExeName : altExeNames) {
-        MenuItem item = new MenuItem(altExeName);
-        item.setOnAction(new EventHandler<ActionEvent>() {
-          @Override
-          public void handle(ActionEvent event) {
-            onPlay(altExeName, null);
-          }
-        });
-        playBtn.getItems().add(item);
+        items.add(new LaunchConfiguration(altExeName, false, altExeName, null));
 
         if (isCameraModeSupported(altExeName)) {
-          item = new MenuItem(altExeName + " [Camera Mode]");
-          item.setOnAction(new EventHandler<ActionEvent>() {
-            @Override
-            public void handle(ActionEvent event) {
-              onPlay(altExeName, "cameraMode");
-            }
-          });
-          playBtn.getItems().add(item);
+          items.add(new LaunchConfiguration(altExeName + " [Camera Mode]", false, altExeName, "cameraMode"));
         }
       }
     }
     else if (game.isFpGame()) {
-      GameEmulatorRepresentation gameEmulator = client.getEmulatorService().getGameEmulator(game.getEmulatorId());
+      gameEmulator = client.getEmulatorService().getGameEmulator(game.getEmulatorId());
       if (gameEmulator != null) {
         List<String> altExeNames = client.getEmulatorService().getAltExeNames(gameEmulator.getId());
         for (String altExeName : altExeNames) {
-          MenuItem item = new MenuItem(altExeName);
-          item.setOnAction(new EventHandler<ActionEvent>() {
-            @Override
-            public void handle(ActionEvent event) {
-              onPlay(altExeName, null);
-            }
-          });
-          playBtn.getItems().add(item);
+          items.add(new LaunchConfiguration(altExeName, false, altExeName, null));
         }
       }
+    }
+
+    launchCombo.setItems(FXCollections.observableList(items));
+    launchCombo.setDisable(items.isEmpty());
+    launchBtn.setDisable(items.isEmpty());
+
+    if (!items.isEmpty()) {
+      this.launchCombo.valueProperty().removeListener(this);
+      launchCombo.getSelectionModel().select(0);
+      UISettings uiSettings = client.getPreferenceService().getJsonPreference(PreferenceNames.UI_SETTINGS, UISettings.class);
+      LaunchConfiguration launchConfiguration = uiSettings.getLaunchConfiguration();
+      if (launchConfiguration != null) {
+        launchCombo.setValue(launchConfiguration);
+      }
+      this.launchCombo.valueProperty().addListener(this);
     }
   }
 
@@ -122,11 +134,8 @@ public class PlayButtonController implements Initializable {
   }
 
   public void setDisable(boolean b) {
-    playBtn.setDisable(b);
-  }
-
-  public void onPlay(String altExe, String option) {
-    onPlay(game, altExe, option);
+    launchBtn.setDisable(b);
+    launchCombo.setDisable(b);
   }
 
   public static void onPlay(GameRepresentation game, String altExe, String option) {
@@ -151,6 +160,15 @@ public class PlayButtonController implements Initializable {
         }
         client.getGameService().playGame(game.getId(), altExe, option);
       }
+    }
+  }
+
+  @Override
+  public void changed(ObservableValue<? extends LaunchConfiguration> observable, LaunchConfiguration oldValue, LaunchConfiguration newValue) {
+    UISettings uiSettings = client.getPreferenceService().getJsonPreference(PreferenceNames.UI_SETTINGS, UISettings.class);
+    if(newValue != null) {
+      uiSettings.setLaunchConfiguration(newValue);
+      client.getPreferenceService().setJsonPreference(uiSettings);
     }
   }
 }